+++ packages/aws-appsync/src/link/complex/complex-object-link-uploader.ts
@@ -16,7 +16,6 @@ export default (fileField, { credentials }) => {
         mimeType: ContentType,
         localUri: Body,
     } = fileField;
-
     const s3 = new S3({
         credentials,
         region,