+++ packages/aws-appsync/src/link/subscription-handshake-link.ts
@@ -8,10 +8,15 @@
  */
 import { ApolloLink, Observable, Operation, FetchResult } from "apollo-link";
 
+import { rootLogger } from "../utils";
 import * as Paho from '../vendor/paho-mqtt';
 import { ApolloError } from "apollo-client";
 import { FieldNode } from "graphql";
 import { getMainDefinition } from "apollo-utilities";
+import { PERMANENT_ERROR_KEY } from "./retry-link";
+
+const logger = rootLogger.extend('subscriptions');
+const mqttLogger = logger.extend('mqtt');
 
 type SubscriptionExtension = {
     mqttConnections: MqttConnectionInfo[],
@@ -62,13 +67,13 @@ export class SubscriptionHandshakeLink extends ApolloLink {
             } = { subscription: { newSubscriptions: {}, mqttConnections: [] } },
             errors = [],
         }: {
-                extensions?: {
-                    subscription: SubscriptionExtension
-                },
-                errors: any[]
-            } = subsInfo;
+            extensions?: {
+                subscription: SubscriptionExtension
+            },
+            errors: any[]
+        } = subsInfo;
 
-        if (errors.length) {
+        if (errors && errors.length) {
             return new Observable(observer => {
                 observer.error(new ApolloError({
                     errorMessage: 'Error during subscription handshake',
@@ -161,12 +166,12 @@ export class SubscriptionHandshakeLink extends ApolloLink {
         const { client: clientId, url, topics } = connectionInfo;
         const client: any = new Paho.Client(url, clientId);
 
-        // client.trace = console.log.bind(null, clientId);
+        client.trace = mqttLogger.bind(null, clientId);
         client.onConnectionLost = ({ errorCode, ...args }) => {
             if (errorCode !== 0) {
                 topics.forEach(t => {
                     if (this.topicObservers.has(t)) {
-                        this.topicObservers.get(t).forEach(observer => observer.error(args));
+                        this.topicObservers.get(t).forEach(observer => observer.error({ ...args, [PERMANENT_ERROR_KEY]: true }));
                     }
                 });
             }
@@ -224,6 +229,8 @@ export class SubscriptionHandshakeLink extends ApolloLink {
             parsedMessage.data || {}
         );
 
+        logger('Message received', { data, topic, observers });
+
         observers.forEach(observer => {
             try {
                 observer.next({
@@ -231,7 +238,7 @@ export class SubscriptionHandshakeLink extends ApolloLink {
                     ...{ data },
                 })
             } catch (err) {
-                // console.error(err);
+                logger(err);
             }
         });
     }