+++ packages/aws-appsync-react/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.2.6"></a>
+## [1.2.6](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.2.5...aws-appsync-react@1.2.6) (2019-01-11)
+
+
+
+
+**Note:** Version bump only for package aws-appsync-react
+
+<a name="1.2.5"></a>
+## [1.2.5](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.2.4...aws-appsync-react@1.2.5) (2018-12-12)
+
+
+
+
+**Note:** Version bump only for package aws-appsync-react
+
+<a name="1.2.4"></a>
+## [1.2.4](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.2.3...aws-appsync-react@1.2.4) (2018-12-07)
+
+
+
+
+**Note:** Version bump only for package aws-appsync-react
+
+<a name="1.2.3"></a>
+## [1.2.3](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.2.2...aws-appsync-react@1.2.3) (2018-12-04)
+
+
+
+
+**Note:** Version bump only for package aws-appsync-react
+
 <a name="1.2.2"></a>
 ## [1.2.2](https://github.com/awslabs/aws-mobile-appsync-sdk-js/compare/aws-appsync-react@1.2.1...aws-appsync-react@1.2.2) (2018-11-21)